removed vestigial for loop in frame constructor
the first for loop was originally used to get these types into an array, no longer needed as the original pairs are not used any further, in the future all of the inner constructors should be based on this. There is also the issue of these now 2 back-to-back loops, which ideally will become one loop of simple in-line syntax.